Added and enabled Test_VARCHAR_MAX_BCP.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@37253 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/test/dbapi_unit_test.cpp
@@ -767,8 +767,8 @@ void CDBAPIUnitTest::Test_Unicode(void)
 void CDBAPIUnitTest::Test_VARCHAR_MAX(void)
 {
     string sql;
-    // const string table_name = "#test_varchar_max_table";
-    const string table_name = "DBAPI_Sample..test_varchar_max_table";
+    const string table_name = "#test_varchar_max_table";
+    // const string table_name = "DBAPI_Sample..test_varchar_max_table";
     // const string msg(32000, 'Z');
     const string msg(8001, 'Z');
 
@@ -787,7 +787,7 @@ void CDBAPIUnitTest::Test_VARCHAR_MAX(void)
         }
 
         // SQL value injection technique ... 
-        {
+        if (true) {
             // Clean table ...
             {
                 sql = "DELETE FROM " + table_name;
@@ -861,6 +861,89 @@ void CDBAPIUnitTest::Test_VARCHAR_MAX(void)
             }
         }
 
+    }
+    catch(const CDB_Exception& ex) {
+        DBAPI_BOOST_FAIL(ex);
+    }
+    catch(const CException& ex) {
+        DBAPI_BOOST_FAIL(ex);
+    }
+}
+
+
+///////////////////////////////////////////////////////////////////////////////
+void CDBAPIUnitTest::Test_VARCHAR_MAX_BCP(void)
+{
+    string sql;
+    const string table_name = "#test_varchar_max_bcp_table";
+    // const string table_name = "DBAPI_Sample..test_varchar_max_bcp_table";
+    // const string msg(32000, 'Z');
+    const string msg(8001, 'Z');
+
+    try {
+        auto_ptr<IStatement> auto_stmt( m_Conn->GetStatement() );
+
+        // Create table ...
+        if (table_name[0] =='#') {
+            sql =
+                "CREATE TABLE " + table_name + " ( \n"
+                "   id NUMERIC IDENTITY NOT NULL, \n"
+                "   vc_max VARCHAR(MAX) NULL" 
+                ") \n";
+
+            auto_stmt->ExecuteUpdate( sql );
+        }
+
+        if (true) {
+            const CVariant vc_max_value(msg);
+
+            // Clean table ...
+            {
+                sql = "DELETE FROM " + table_name;
+
+                auto_stmt->ExecuteUpdate( sql );
+            }
+
+            // Insert data into the table ...
+            {
+                auto_ptr<IBulkInsert> bi(
+                    m_Conn->GetBulkInsert(table_name)
+                    );
+                CVariant col1(eDB_Int);
+                // CVariant col2(eDB_VarChar);
+                CVariant col2(eDB_Text);
+
+                bi->Bind(1, &col1);
+                bi->Bind(2, &col2);
+
+                col1 = 1;
+                // col2 = msg;
+                col2.Append(msg.c_str(), msg.size());
+
+                bi->AddRow();
+                bi->Complete();
+            }
+
+            // Actual check ...
+            {
+                sql = "SELECT vc_max FROM " + table_name + " ORDER BY id";
+
+                auto_stmt->SendSql( sql );
+                while( auto_stmt->HasMoreResults() ) {
+                    if( auto_stmt->HasRows() ) {
+                        auto_ptr<IResultSet> rs(auto_stmt->GetResultSet());
+                        BOOST_CHECK( rs.get() != NULL );
+
+                        rs->BindBlobToVariant(true);
+                        BOOST_CHECK( rs->Next() );
+                        BOOST_CHECK_EQUAL(msg.size(), rs->GetVariant(1).GetBlobSize());
+                        // const string value = rs->GetVariant(1).GetString();
+                        // BOOST_CHECK_EQUAL(value.size(), msg.size());
+                        // BOOST_CHECK_EQUAL(value, msg);
+                    }
+                }
+            }
+        }
     }
     catch(const CDB_Exception& ex) {
         DBAPI_BOOST_FAIL(ex);
@@ -13317,9 +13400,19 @@ CDBAPITestSuite::CDBAPITestSuite(const CTestArguments& args)
 
     add(tc_init);
 
-    // if (args.GetServerType() == CTestArguments::eMsSql2005) {
+    // if (args.GetServerType() == CTestArguments::eMsSql2005
     if (false) {
         tc = BOOST_CLASS_TEST_CASE(&CDBAPIUnitTest::Test_VARCHAR_MAX,
+                DBAPIInstance);
+        tc->depends_on(tc_init);
+        add(tc);
+    }
+
+    if (args.IsBCPAvailable()
+        && args.GetServerType() == CTestArguments::eMsSql2005
+        && args.GetDriverName() != dblib_driver
+        ) {
+        tc = BOOST_CLASS_TEST_CASE(&CDBAPIUnitTest::Test_VARCHAR_MAX_BCP,
                 DBAPIInstance);
         tc->depends_on(tc_init);
         add(tc);
